fix mp3 statusbar on postdel, store genre and artist in racedata (rm -fR pzs-ng/site/* needed - i'll add more stuff to store in a later commit), fix error re: ul/dl-speedtesting, fix incomplete-link in postdel. And probably something else.

+++ trunk/sitebot/src/showlog.c
@@ -329,13 +329,13 @@ void show_newdirs(const char *pattern)
 
 if (!matchpath(group_dirs, buffer.dirname)) {
 #ifndef USE_GLFTPD132
-			printf("%d|%u|%d|%d|%d|%lld|",
+			printf("%d|%u|%d|%d|%d|%.0f|",
 				buffer.status, (unsigned int)buffer.uptime, buffer.uploader, buffer.group,
-				buffer.files, buffer.bytes/1024);
+				buffer.files, (float)buffer.bytes/1024.);
 #else
-			printf("%d|%u|%d|%d|%d|%ld|",
+			printf("%d|%u|%d|%d|%d|%.0f|",
 				buffer.status, (unsigned int)buffer.uptime, buffer.uploader, buffer.group,
-				buffer.files, buffer.bytes/1024);
+				buffer.files, (float)buffer.bytes/1024.);
 #endif
 			puts(buffer.dirname);
 			i++;
